Add external MQTT broker bridge mode (v3.4.0)
- Extract shared logic into MQTTBase abstract class (mqttBase.ts)
- Add MQTTBridge: connects as MQTT client to an external broker instead
  of running a built-in TCP server
- Hostname-based device naming: devices are identified and renamed via
  the Hostname field in tele/+/STATE or stat/+/STATUS6 (MqttClient)
- Admin UI: new "External MQTT broker" section with URL, user, password
- i18n: translations for all 12 languages
- Config: useExternalBroker, externalBrokerUrl/User/Password fields
- Fix adapter connection status: info.connection now reflects server
  listening state (server mode) or broker URL (bridge mode) instead of
  the device list, so the indicator is green immediately on startup
