+++ admin/i18n/it.json
@@ -60,6 +60,13 @@
     "Use different topic names for set and get:": "Usa diversi nomi di argomenti per impostare e ottenere",
     "User:": "Utente",
     "WARNING! Check README.md!": "AVVERTIMENTO! Controlla README.md!",
+    "External MQTT broker": "Broker MQTT esterno",
+    "Use external MQTT broker": "Usa broker MQTT esterno",
+    "Disables the built-in MQTT server and connects as client to an external broker": "Disabilita il server MQTT integrato e si connette come client a un broker esterno",
+    "External broker URL": "URL del broker esterno",
+    "e.g. mqtt://192.168.1.100:1883 or mqtts://broker:8883": "es. mqtt://192.168.1.100:1883 o mqtts://broker:8883",
+    "External broker user": "Utente del broker esterno",
+    "External broker password": "Password del broker esterno",
     "chars": "caratteri",
     "connected": "collegato",
     "ms": "Signorina"
