+++ admin/i18n/pt.json
@@ -60,6 +60,13 @@
     "Use different topic names for set and get:": "Use nomes de tópicos diferentes para definir e obter",
     "User:": "Utilizador",
     "WARNING! Check README.md!": "AVISO! Verifique README.md!",
+    "External MQTT broker": "Broker MQTT externo",
+    "Use external MQTT broker": "Usar broker MQTT externo",
+    "Disables the built-in MQTT server and connects as client to an external broker": "Desativa o servidor MQTT integrado e conecta como cliente a um broker externo",
+    "External broker URL": "URL do broker externo",
+    "e.g. mqtt://192.168.1.100:1883 or mqtts://broker:8883": "ex. mqtt://192.168.1.100:1883 ou mqtts://broker:8883",
+    "External broker user": "Usuário do broker externo",
+    "External broker password": "Senha do broker externo",
     "chars": "caracteres",
     "connected": "conectado",
     "ms": "Senhora"
