+++ admin/jsonConfig.json
@@ -12,11 +12,68 @@
             "lg": 12,
             "xl": 12
         },
+        "_headerExternal": {
+            "newLine": true,
+            "type": "header",
+            "text": "External MQTT broker",
+            "size": 3,
+            "xs": 12,
+            "sm": 12,
+            "md": 12,
+            "lg": 12,
+            "xl": 12
+        },
+        "useExternalBroker": {
+            "newLine": true,
+            "type": "checkbox",
+            "label": "Use external MQTT broker",
+            "help": "Disables the built-in MQTT server and connects as client to an external broker",
+            "xs": 12,
+            "sm": 12,
+            "md": 12,
+            "lg": 12,
+            "xl": 12
+        },
+        "externalBrokerUrl": {
+            "newLine": true,
+            "type": "text",
+            "label": "External broker URL",
+            "help": "e.g. mqtt://192.168.1.100:1883 or mqtts://broker:8883",
+            "hidden": "!data.useExternalBroker",
+            "xs": 12,
+            "sm": 12,
+            "md": 12,
+            "lg": 8,
+            "xl": 8
+        },
+        "externalBrokerUser": {
+            "newLine": true,
+            "type": "text",
+            "label": "External broker user",
+            "hidden": "!data.useExternalBroker",
+            "xs": 12,
+            "sm": 12,
+            "md": 6,
+            "lg": 4,
+            "xl": 4
+        },
+        "externalBrokerPassword": {
+            "type": "password",
+            "label": "External broker password",
+            "repeat": true,
+            "hidden": "!data.useExternalBroker",
+            "xs": 12,
+            "sm": 12,
+            "md": 6,
+            "lg": 4,
+            "xl": 4
+        },
         "bind": {
             "newLine": true,
             "type": "ip",
             "label": "IP",
             "listenOnAllPorts": true,
+            "hidden": "!!data.useExternalBroker",
             "xs": 12,
             "sm": 12,
             "md": 6,
@@ -28,6 +85,7 @@
             "type": "number",
             "min": 1,
             "max": 65535,
+            "hidden": "!!data.useExternalBroker",
             "xs": 12,
             "sm": 12,
             "md": 6,
@@ -40,6 +98,7 @@
             "type": "number",
             "min": 1,
             "max": 99999,
+            "hidden": "!!data.useExternalBroker",
             "xs": 12,
             "sm": 12,
             "md": 6,
@@ -51,6 +110,7 @@
             "type": "header",
             "text": "Authentication settings",
             "size": 3,
+            "hidden": "!!data.useExternalBroker",
             "xs": 12,
             "sm": 12,
             "md": 12,
@@ -61,6 +121,7 @@
             "newLine": true,
             "type": "text",
             "label": "User",
+            "hidden": "!!data.useExternalBroker",
             "xs": 12,
             "sm": 12,
             "md": 6,
@@ -75,7 +136,8 @@
             "xl": 4,
             "type": "password",
             "label": "Password",
-            "repeat": true
+            "repeat": true,
+            "hidden": "!!data.useExternalBroker"
         },
         "_header3": {
             "newLine": true,
@@ -184,6 +246,7 @@
             "help": "Time in minutes to store client information",
             "min": 1,
             "max": 99999,
+            "hidden": "!!data.useExternalBroker",
             "xs": 12,
             "sm": 12,
             "md": 6,
@@ -208,6 +271,7 @@
             "help": "Interval in milliseconds for message retransmission",
             "min": 100,
             "max": 99999,
+            "hidden": "!!data.useExternalBroker",
             "xs": 12,
             "sm": 12,
             "md": 6,
@@ -220,6 +284,7 @@
             "help": "Maximum number of retransmission attempts",
             "min": 1,
             "max": 100,
+            "hidden": "!!data.useExternalBroker",
             "xs": 12,
             "sm": 12,
             "md": 6,
