+++ src/main.ts
@@ -6,12 +6,14 @@
  *      MIT License
  */
 import { Adapter, type AdapterOptions } from '@iobroker/adapter-core'; // Get common this utils
-import Server from './lib/server';
+import MQTTServer from './lib/server';
+import MQTTBridge from './lib/bridge';
+import type MQTTBase from './lib/mqttBase';
 import type { SonoffAdapterConfig } from './types';
 
 export class SonoffAdapter extends Adapter {
     declare config: SonoffAdapterConfig;
-    server: Server | null = null;
+    server: MQTTBase | null = null;
 
     public constructor(options: Partial<AdapterOptions> = {}) {
         super({
@@ -40,6 +42,19 @@ export class SonoffAdapter extends Adapter {
     }
 
     private async main(): Promise<void> {
+        // Reset any pending (ack=false) states before subscribing and starting the server.
+        // On restart, ioBroker may re-deliver unacknowledged states to the adapter, which
+        // would cause the server to send commands to devices (e.g. opening a garage door).
+        // Acknowledging them here ensures they are treated as read-only current values.
+        const currentStates = await this.getStatesAsync('*');
+        if (currentStates) {
+            for (const [id, state] of Object.entries(currentStates)) {
+                if (state && !state.ack) {
+                    await this.setForeignStateAsync(id, state.val, true);
+                }
+            }
+        }
+
         // subscribe for all own variables
         this.subscribeStates('*');
 
@@ -53,7 +68,11 @@ export class SonoffAdapter extends Adapter {
             }
         }
 
-        this.server = new Server(this as ioBroker.Adapter);
+        if (this.config.useExternalBroker) {
+            this.server = new MQTTBridge(this as ioBroker.Adapter);
+        } else {
+            this.server = new MQTTServer(this as ioBroker.Adapter);
+        }
     }
 }
